+++ Code/Engine/Foundation/Platform/UWP/FormatString_Platform.h
@@ -1,0 +1,2 @@
+// redirect to shared implementation
+#include <Foundation/Platform/Win/FormatString_Platform.h>
