+++ Code/Engine/Foundation/Strings/Implementation/FormatString.cpp
@@ -400,49 +400,3 @@ ezStringView ezArgSensitive::BuildString_SensitiveUserData_Hash(char* szTmp, ezU
 
   return szTmp;
 }
-
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
-#  include <Foundation/Platform/Win/Utils/IncludeWindows.h>
-
-ezStringView BuildString(char* szTmp, ezUInt32 uiLength, const ezArgErrorCode& arg)
-{
-  LPVOID lpMsgBuf = nullptr;
-  if (FormatMessageW(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, nullptr, arg.m_ErrorCode,
-        MAKELANGID(LANG_ENGLISH, SUBLANG_ENGLISH_US), (LPWSTR)&lpMsgBuf, 0, nullptr) == 0)
-  {
-    DWORD err = GetLastError();
-    ezStringUtils::snprintf(szTmp, uiLength, "%i (FormatMessageW failed with error code %i)", arg.m_ErrorCode, err);
-    return ezStringView(szTmp);
-  }
-
-  LPWSTR pCRLF = wcschr((LPWSTR)lpMsgBuf, L'\r');
-  if (pCRLF != nullptr)
-  {
-    // remove the \r\n that FormatMessageW always appends
-    *pCRLF = L'\0';
-  }
-
-  // we need a bigger boat
-  static thread_local char FullMessage[256];
-
-  ezStringUtils::snprintf(FullMessage, EZ_ARRAY_SIZE(FullMessage), "%i (\"%s\")", arg.m_ErrorCode, ezStringUtf8((LPWSTR)lpMsgBuf).GetData());
-  LocalFree(lpMsgBuf);
-  return ezStringView(FullMessage);
-}
-#else
-ezStringView BuildString(char* szTmp, ezUInt32 uiLength, const ezArgErrorCode& arg)
-{
-  return "NOT_SUPPORTED";
-}
-#endif
-
-#if EZ_ENABLED(EZ_PLATFORM_LINUX) || EZ_ENABLED(EZ_PLATFORM_ANDROID)
-#  include <string.h>
-
-ezStringView BuildString(char* szTmp, ezUInt32 uiLength, const ezArgErrno& arg)
-{
-  const char* szErrorMsg = std::strerror(arg.m_iErrno);
-  ezStringUtils::snprintf(szTmp, uiLength, "%i (\"%s\")", arg.m_iErrno, szErrorMsg);
-  return ezStringView(szTmp);
-}
-#endif
