+++ Code/Engine/Foundation/Strings/Implementation/FormatStringArgs.h
@@ -133,38 +133,6 @@ struct ezArgFileSize : public ezArgHumanReadable
   const char* const m_ByteSuffixes[6] = {"B", "KB", "MB", "GB", "TB", "PB"};
 };
 
-/// \brief Converts a windows HRESULT into an error code and a human-readable error message.
-/// Pass in `GetLastError()` function or an HRESULT from another error source. Be careful when printing multiple values, a function could clear `GetLastError` as a side-effect so it is best to store it in a temp variable before printing a complex error message.
-/// \sa https://learn.microsoft.com/en-gb/windows/win32/api/errhandlingapi/nf-errhandlingapi-getlasterror
-struct ezArgErrorCode
-{
-  inline explicit ezArgErrorCode(ezUInt32 uiErrorCode)
-    : m_ErrorCode(uiErrorCode)
-  {
-  }
-
-  ezUInt32 m_ErrorCode;
-};
-EZ_FOUNDATION_DLL ezStringView BuildString(char* szTmp, ezUInt32 uiLength, const ezArgErrorCode& arg);
-
-
-#if EZ_ENABLED(EZ_PLATFORM_LINUX) || EZ_ENABLED(EZ_PLATFORM_ANDROID)
-/// \brief Many Linux APIs will fill out error on failure. This converts the error into an error code and a human-readable error message.
-/// Pass in the linux `errno` symbol. Be careful when printing multiple values, a function could clear `errno` as a side-effect so it is best to store it in a temp variable before printing a complex error message.
-/// You may have to include #include <errno.h> use this.
-/// \sa https://man7.org/linux/man-pages/man3/errno.3.html
-struct ezArgErrno
-{
-  inline explicit ezArgErrno(ezInt32 iErrno)
-    : m_iErrno(iErrno)
-  {
-  }
-
-  ezInt32 m_iErrno;
-};
-EZ_FOUNDATION_DLL ezStringView BuildString(char* szTmp, ezUInt32 uiLength, const ezArgErrno& arg);
-#endif
-
 /// \brief Wraps a string that may contain sensitive information, such as user file paths.
 ///
 /// The application can specify a function to scramble this type of information. By default no such function is set.
@@ -230,3 +198,6 @@ EZ_ALWAYS_INLINE ezStringView BuildString(char* szTmp, ezUInt32 uiLength, unsign
 }
 
 #endif
+
+// add platform specific formatters
+#include <FormatString_Platform.h>
